+++ backend/canisters/group_index/impl/src/updates/upgrade_local_group_index_canister_wasm.rs
@@ -4,7 +4,6 @@ use canister_api_macros::proposal;
 use canister_tracing_macros::trace;
 use group_index_canister::upgrade_local_group_index_canister_wasm::*;
 use group_index_canister::ChildCanisterType;
-use std::collections::HashSet;
 use tracing::info;
 use types::{BuildVersion, CanisterWasm, UpgradeChunkedCanisterWasmResponse::*};
 use utils::canister::should_perform_upgrade;
@@ -42,18 +41,13 @@ fn upgrade_local_group_index_canister_wasm_impl(args: Args, state: &mut RuntimeS
             .set(ChildCanisterType::LocalGroupIndex, CanisterWasm { version, module: wasm });
 
         let filter = args.filter.unwrap_or_default();
-        let include: HashSet<_> = filter.include.into_iter().collect();
-        let include_all = include.is_empty();
-        let exclude: HashSet<_> = filter.exclude.into_iter().collect();
 
         for canister_id in state
             .data
             .local_index_map
             .iter()
-            .filter(|(_, i)| should_perform_upgrade(i.wasm_version(), version, state.data.test_mode))
+            .filter(|(id, index)| should_perform_upgrade(**id, index.wasm_version(), version, &filter, state.data.test_mode))
             .map(|(c, _)| *c)
-            .filter(|c| include_all || include.contains(c))
-            .filter(|c| !exclude.contains(c))
         {
             state.data.canisters_requiring_upgrade.enqueue(canister_id, false);
         }