+++ backend/libraries/types/src/build_version.rs
@@ -5,7 +5,7 @@ use std::str::FromStr;
 use ts_export::ts_export;
 
 #[ts_export]
-#[derive(CandidType, Serialize, Deserialize, Clone, Copy, Debug, Default, Ord, PartialOrd, Eq, PartialEq)]
+#[derive(CandidType, Serialize, Deserialize, Clone, Copy, Debug, Default, Ord, PartialOrd, Eq, PartialEq, Hash)]
 pub struct BuildVersion {
     pub major: u32,
     pub minor: u32,